requirements.txt
Pillow==9.2.0
pymediainfo==5.1.0
